refactor(landing): colocate landing page logic into route-level private folders
- move components to _components
- move data to _data
- move types to _types
- remove unused global components
- improve code organization and maintainability
