+++ apps/web/app/(full-layout)/(home)/_components/index.ts
@@ -1,0 +1,4 @@
+export { ContactSection } from './contact-section';
+export { HeroSection } from './hero-section';
+export { PersonalizedSection } from './personalized-section';
+export { RoadmapSection } from './roadmap-section';
