+++ apps/web/components/shared/mask-background.tsx
