+++ apps/web/components/shared/rainbow-bar.tsx
